Improved readability + added documentation/description
